AI Agent & RAG/requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.10

# LLM Providers
openai>=1.10.0
anthropic>=0.18.0
transformers>=4.36.0
torch>=2.1.0

# Embeddings
sentence-transformers>=2.2.2
chromadb>=0.4.22
faiss-cpu>=1.7.4

# Vector stores and databases
pinecone-client>=2.2.0
weaviate-client>=4.4.0

# Document processing
pypdf>=3.17.0
unstructured>=0.11.0
pdf2image>=1.16.3
pytesseract>=0.3.10

# Text processing
nltk>=3.8.1
spacy>=3.7.2
beautifulsoup4>=4.12.2

# Utilities
python-dotenv>=1.0.0
pandas>=2.1.0
numpy>=1.24.0
tiktoken>=0.5.0

# Reranking
cohere>=4.37.0

# Tools and computation
numpy>=1.24.0
pandas>=2.1.0

# API and web
fastapi>=0.104.0
uvicorn>=0.24.0
requests>=2.31.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
